pyproject.toml
[tool.poetry]
name = "py_opengl"
version = "0.1.0"
description = ""
authors = ["nalwayv"]

[tool.poetry.dependencies]
python = "^3.10"
glfw = "^2.4.0"
PyOpenGL = "^3.1.5"
loguru = "^0.5.3"
Pillow = "^9.0.0"

[tool.poetry.scripts]
main = "py_opengl.main:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
